+++ tests/macro_tests.rs
@@ -9,7 +9,7 @@ extern crate frunk;
 fn use_frunk_macros() {
     let h1 = hlist![1i32, 2u32];
     let h2 = hlist!["cool", ...h1];
-    let hlist_pat![a, ...bs]: Hlist![&'static str, i32, ...Hlist![u32]] = h2;
+    let hlist_pat![a, ...bs]: HList![&'static str, i32, ...HList![u32]] = h2;
     assert_eq!(a, "cool");
     assert_eq!(bs, h1);
 }